// src/api/generated/model/hotSpotArea.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export interface HotSpotArea {
  id?: string;
  shape?: string;
  coordinates?: string;
  isCorrect?: boolean;
  feedback?: string;
  label?: string;
}