// src/api/generated/model/hotSpotOptions.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { HotSpotArea } from "./hotSpotArea";

export interface HotSpotOptions {
  backgroundImageUrl?: string;
  hotSpots?: HotSpotArea[];
  selectionType?: string;
}